+++ data/bone_settings.lua
@@ -1,0 +1,72 @@
+return {
+    ["HEAD"] = {
+        label = "Head",
+        severity = 0
+    },
+    ["NECK"] = {
+        label = "Neck",
+        severity = 0
+    },
+    ["SPINE"] = {
+        label = "Spine",
+        severity = 0,
+        causeLimp = true
+    },
+    ["UPPER_BODY"] = {
+        label = "Upper Body",
+        severity = 0
+    },
+    ["LOWER_BODY"] = {
+        label = "Lower Body",
+        severity = 0,
+        causeLimp = true
+    },
+    ["LARM"] = {
+        label = "Left Arm",
+        severity = 0,
+        canUseTourniquet = true,
+    },
+    ["LHAND"] = {
+        label = "Left Hand",
+        severity = 0
+    },
+    ["LFINGER"] = {
+        label = "Left Hand Fingers",
+        severity = 0
+    },
+    ["LLEG"] = {
+        label = "Left Leg",
+        severity = 0,
+        causeLimp = true,
+        canUseTourniquet = true,
+    },
+    ["LFOOT"] = {
+        label = "Left Foot",
+        severity = 0,
+        causeLimp = true
+    },
+    ["RARM"] = {
+        label = "Right Arm",
+        severity = 0,
+        canUseTourniquet = true,
+    },
+    ["RHAND"] = {
+        label = "Right Hand",
+        severity = 0
+    },
+    ["RFINGER"] = {
+        label = "Right Hand Fingers",
+        severity = 0
+    },
+    ["RLEG"] = {
+        label = "Right Leg",
+        severity = 0,
+        causeLimp = true,
+        canUseTourniquet = true,
+    },
+    ["RFOOT"] = {
+        label = "Right Foot",
+        severity = 0,
+        causeLimp = true
+    }
+}
