+++ data/bones.lua
@@ -1,0 +1,58 @@
+return {
+    [0] = "NONE",
+    [31085] = "HEAD",
+    [31086] = "HEAD",
+    [39317] = "NECK",
+    [57597] = "SPINE",
+    [23553] = "SPINE",
+    [24816] = "SPINE",
+    [24817] = "SPINE",
+    [24818] = "SPINE",
+    [10706] = "UPPER_BODY",
+    [64729] = "UPPER_BODY",
+    [11816] = "LOWER_BODY",
+    [45509] = "LARM",
+    [61163] = "LARM",
+    [18905] = "LHAND",
+    [4089] = "LFINGER",
+    [4090] = "LFINGER",
+    [4137] = "LFINGER",
+    [4138] = "LFINGER",
+    [4153] = "LFINGER",
+    [4154] = "LFINGER",
+    [4169] = "LFINGER",
+    [4170] = "LFINGER",
+    [4185] = "LFINGER",
+    [4186] = "LFINGER",
+    [26610] = "LFINGER",
+    [26611] = "LFINGER",
+    [26612] = "LFINGER",
+    [26613] = "LFINGER",
+    [26614] = "LFINGER",
+    [58271] = "LLEG",
+    [63931] = "LLEG",
+    [2108] = "LFOOT",
+    [14201] = "LFOOT",
+    [40269] = "RARM",
+    [28252] = "RARM",
+    [57005] = "RHAND",
+    [58866] = "RFINGER",
+    [58867] = "RFINGER",
+    [58868] = "RFINGER",
+    [58869] = "RFINGER",
+    [58870] = "RFINGER",
+    [64016] = "RFINGER",
+    [64017] = "RFINGER",
+    [64064] = "RFINGER",
+    [64065] = "RFINGER",
+    [64080] = "RFINGER",
+    [64081] = "RFINGER",
+    [64096] = "RFINGER",
+    [64097] = "RFINGER",
+    [64112] = "RFINGER",
+    [64113] = "RFINGER",
+    [36864] = "RLEG",
+    [51826] = "RLEG",
+    [20781] = "RFOOT",
+    [52301] = "RFOOT"
+}
