+++ fxmanifest.lua
@@ -14,18 +14,15 @@ dependencies {
     "ND_Core"
 }
 
+files {
+    "data/**"
+}
+
 shared_scripts {
     "@ox_lib/init.lua",
-    "@ND_Core/init.lua",
-    "data.lua"
+    "@ND_Core/init.lua"
 }
-server_scripts {
-    "server/main.lua",
-    "server/body.lua",
-    "server/items/**"
-}
+
 client_scripts {
-    "client/body.lua",
-    "client/job.lua",
-    "client/items/**"
+    "client/**"
 }
